Move 'troubleshooting' command to 'diagnose'
