--- cmd/agent/subcommands/troubleshooting/command.go
@@ -1,116 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-// Package troubleshooting implements 'agent troubleshooting'.
-package troubleshooting
-
-import (
-	"fmt"
-	"os"
-
-	"github.com/spf13/cobra"
-	"go.uber.org/fx"
-
-	"github.com/DataDog/datadog-agent/cmd/agent/command"
-	"github.com/DataDog/datadog-agent/comp/core"
-	"github.com/DataDog/datadog-agent/comp/core/config"
-	"github.com/DataDog/datadog-agent/comp/core/log"
-	"github.com/DataDog/datadog-agent/pkg/api/util"
-	pkgconfig "github.com/DataDog/datadog-agent/pkg/config"
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
-)
-
-const (
-	metadataEndpoint = "/agent/metadata/"
-)
-
-// cliParams are the command-line arguments for this subcommand
-type cliParams struct {
-	*command.GlobalParams
-
-	// payloadName is the name of the payload to display
-	payloadName string
-}
-
-// Commands returns a slice of subcommands for the 'agent' command.
-func Commands(globalParams *command.GlobalParams) []*cobra.Command {
-	cliParams := &cliParams{
-		GlobalParams: globalParams,
-	}
-	payloadV5Cmd := &cobra.Command{
-		Use:   "metadata_v5",
-		Short: "Print the metadata payload for the agent.",
-		Long: `
-This command print the V5 metadata payload for the Agent. This payload is used to populate the infra list and host map in Datadog. It's called 'V5' because it's the same payload sent since Agent V5. This payload is mandatory in order to create a new host in Datadog.`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			cliParams.payloadName = "v5"
-			return fxutil.OneShot(printPayload,
-				fx.Supply(cliParams),
-				fx.Supply(core.BundleParams{
-					ConfFilePath:      globalParams.ConfFilePath,
-					ConfigLoadSecrets: false,
-				}.LogForOneShot("CORE", "off", true)),
-				core.Bundle,
-			)
-		},
-	}
-
-	payloadInventoriesCmd := &cobra.Command{
-		Use:   "metadata_inventory",
-		Short: "Print the Inventory metadata payload for the agent.",
-		Long: `
-This command print the last Inventory metadata payload sent by the Agent. This payload is used by the 'inventories/sql' product.`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			cliParams.payloadName = "inventory"
-			return fxutil.OneShot(printPayload,
-				fx.Supply(cliParams),
-				fx.Supply(core.BundleParams{
-					ConfFilePath:      globalParams.ConfFilePath,
-					ConfigLoadSecrets: false,
-				}.LogForOneShot("CORE", "off", true)),
-				core.Bundle,
-			)
-		},
-	}
-
-	troubleshootingCmd := &cobra.Command{
-		Use:   "troubleshooting",
-		Short: "Helpers to troubleshoot the Datadog Agent.",
-		Long: `
-This command offers a list of helpers to troubleshoot the Datadog Agent.`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			cmd.Help() //nolint:errcheck
-			os.Exit(0)
-			return nil
-		},
-	}
-	troubleshootingCmd.AddCommand(payloadV5Cmd)
-	troubleshootingCmd.AddCommand(payloadInventoriesCmd)
-
-	return []*cobra.Command{troubleshootingCmd}
-}
-
-func printPayload(log log.Component, config config.Component, cliParams *cliParams) error {
-	if err := util.SetAuthToken(); err != nil {
-		fmt.Println(err)
-		return nil
-	}
-
-	c := util.GetClient(false)
-	ipcAddress, err := pkgconfig.GetIPCAddress()
-	if err != nil {
-		return err
-	}
-	apiConfigURL := fmt.Sprintf("https://%v:%d%s%s",
-		ipcAddress, config.GetInt("cmd_port"), metadataEndpoint, cliParams.payloadName)
-
-	r, err := util.DoGet(c, apiConfigURL, util.CloseConnection)
-	if err != nil {
-		return fmt.Errorf("Could not fetch metadata v5 payload: %s", err)
-	}
-
-	fmt.Println(string(r))
-	return nil
-}
