+++ releasenotes/notes/deprecate-troubleshooting-command-cbc23c04b7c4f37a.yaml
@@ -1,0 +1,5 @@
+---
+upgrade:
+  - |
+    Troubleshooting commands in the Agent CLI have been move to the `diagnose` command. `troubleshooting metadata_v5`
+    command is now `diagnose show-metadata v5` and `troubleshooting metadata_inventory` is `diagnose show-metadata inventory`.
